refactor: rename EmailVerificationToken.spec.js to EmailVerification.spec.js
